Created event_parser.py, minor changes
