+++ simulator/simulator.py
@@ -107,58 +107,3 @@ def init_elements(data, simulator):
         else:
             logging.error("Invalid SCIONElement %s", items[0])
             sys.exit()
-
-
-class EventParser(object):
-    """
-    Command parser for the events conf file
-    """
-    def __init__(self, simulator):
-        """
-        Initialize the Event Parser.
-
-        :param simulator: Instance of simulator class.
-        :type simulator: Simulator
-        """
-        self.simulator = simulator
-
-    def parse(self, command):
-        """
-        Parses the command and adds an event.
-
-        :param command: The command
-        :type command: str
-        """
-        items = command.split()
-        router_name = items[0]
-        event_time = float(items[2])
-        if items[1] == "start":
-            self.start_router(router_name, event_time)
-        elif items[1] == "stop":
-            self.stop_router(router_name, event_time)
-        else:
-            logging.error("Command not recognized")
-
-    def start_router(self, router, time):
-        """
-        Start the router at specified time.
-
-        :param router: The name of the router(E.g., er1-19er1-16)
-        :type router: str
-        :param time: Time at which router is to be started
-        :type time: float
-        """
-        self.simulator.add_event(time, cb=self.simulator.start_element,
-                                 args=(router,))
-
-    def stop_router(self, router, time):
-        """
-        Stop the router at specified time.
-
-        :param router: The name of the router(E.g., er1-19er1-16)
-        :type router: str
-        :param time: Time at which router is to be started
-        :type time: float
-        """
-        self.simulator.add_event(time, cb=self.simulator.stop_element,
-                                 args=(router,))
